Move type map, overwrite test for type map
