+++ include/aerial_autonomy/common/type_map.h
@@ -4,8 +4,14 @@
 // store objects
 #include <unordered_map>
 
+/**
+* @brief Store objects with a common base class
+*
+* @tparam GenericObjectT Common base class for objects stored
+*/
 template <class GenericObjectT> class TypeMap {
   // Map to store the controller to hardware connectors
+  // TODO Gowtham Maybe use shared/unique pointers
   std::unordered_map<std::type_index, GenericObjectT *> object_storage_map_;
 
 public:
